dbeaver/pro#6202 Extract patching-related code to a separate class
